Update helperFns.cpp
A minor aesthetic change to prevent gamers from getting annoyed when they play PacVim. The word 'game' can be used to refer to PacVim as a whole, and it can be used to refer to a single attempt to beat the game without losing all lives. 

However, it cannot be used to describe each stage on its own; there's another word for that: 'level'. 

Hence why now the game should say "LEVEL UP!" each time you beat a map.

A further change that's incoming is to get the game to say "GAME OVER!" when you lose all your lives.